Change to flex layout
