+++ packages/app/src/components/App.css
@@ -37,7 +37,7 @@ body {
     background-color: #aaa;
     grid-gap: 1px;
     grid-template-columns: 250px auto 385px;
-    grid-template-rows: 48px auto 90px auto;
+    grid-template-rows: 48px auto auto auto;
     grid-template-areas:
         'headerbar headerbar headerbar'
         'charttype menubar menubar'
